+++ R/colour.R
@@ -219,49 +219,6 @@ print.color_scheme <- function(x, ...) {
   }
 }
 
-#' Color Ramp
-#'
-#' Provides a color scheme that map a `numeric` vector to colors.
-#' @param x A [`numeric`] vector to be mapped to colors.
-#' @param palette A [`character`] string giving the name of the palette to be
-#'  used (see [color()]).
-#' @param midpoint A [`numeric`] value specifying the midpoint of `x`. If not
-#'  `NULL`, `x` will be rescaled to have the specified `midpoint`.
-#' @param ... Further arguments passed to [color()].
-#' @return A palette function, that when called with a single argument
-#'  (the number of levels, e.g. the `breaks` returned by [hist()])
-#'  returns a vector of colors.
-#' @example inst/examples/ex-ramp.R
-#' @author N. Frerebeau
-#' @family color palettes
-#' @keywords color
-#' @export
-ramp <- function(x, palette, midpoint = NULL, ...) {
-
-  fun <- function(n) {
-
-    if (!is.null(midpoint) && is.numeric(midpoint)) {
-      to <- c(0, 1)
-      from <- range(x, na.rm = TRUE)
-      extent <- 2 * max(abs(from - midpoint))
-      z <- (x - midpoint) / extent * diff(to) + mean(to)
-    } else {
-      z <- (x - min(x)) / (max(x) - min(x))
-    }
-
-    col <- color(palette, ...)
-    ramp <- grDevices::colorRamp(col(n))(z)
-
-    ## Set attributes
-    structure(
-      grDevices::rgb(ramp[, 1], ramp[, 2], ramp[, 3], maxColorValue = 255),
-      breaks = x,
-      class = c("color_scheme")
-    )
-  }
-  return(fun)
-}
-
 #' Information
 #'
 #' Returns information about the available palettes.