+++ NEWS.md
@@ -1,4 +1,6 @@
 # khroma 1.10.0.9000
+## Bugfixes & changes
+*  Fix `ramp()` to return a palette function with argument a vector of values that are mapped to colors.
 
 # khroma 1.10.0
 ## Bugfixes & changes